+++ build.gradle
@@ -4,7 +4,7 @@ plugins { // Declaración de los complementos que se van a utilizar en el proyec
 	id 'io.spring.dependency-management' version '1.1.0' // Complemento para la gestión de dependencias de Spring Boot
 }
 
-group = 'com.test.retirarefectivo'// Identificación del grupo de la aplicación
+group = 'com.bankiko.atm.login'// Identificación del grupo de la aplicación
 version = '0.0.1-SNAPSHOT'// Versión de la aplicación
 sourceCompatibility = '17'// Versión de Java utilizada para compilar el código fuente
