+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'retirarefectivo'
+rootProject.name = 'login'
