+++ src/main/java/com/banckiko/atm/login/LoginApplication.java
@@ -1,0 +1,13 @@
+package com.banckiko.atm.login;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LoginApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LoginApplication.class, args);
+	}
+
+}
