+++ src/main/java/com/banckiko/atm/login/model/registro/RegistroCuentasRequest.java
@@ -1,0 +1,13 @@
+package com.banckiko.atm.login.model.registro;
+
+import lombok.Data;
+
+@Data
+public class RegistroCuentasRequest {
+
+    private String nombre;
+    private Long cuenta;
+    private Integer status;
+
+
+}
