+++ src/main/java/com/banckiko/atm/login/utilidades/Folios.java
@@ -1,0 +1,23 @@
+package com.banckiko.atm.login.utilidades;
+
+import org.springframework.stereotype.Component;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.UUID;
+
+@Component
+public class Folios {
+
+
+    public String folioResponse(){
+        //MSR-160223-ASKDJHG45
+        String ramdomText = UUID.randomUUID().toString().toUpperCase().replace("-","").substring(0,10);
+        Date date = new Date();
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("ddMMyyyy");
+        String fechaActual = simpleDateFormat.format(date);
+        return  "MSR-"+fechaActual+"-"+ramdomText;
+    }
+
+
+}
