--- src/main/java/com/test/retirarefectivo/RetirarefectivoApplication.java
@@ -1,13 +1,0 @@
-package com.test.retirarefectivo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RetirarefectivoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RetirarefectivoApplication.class, args);
-	}
-
-}
