--- src/main/java/com/test/retirarefectivo/model/registro/RegistroCuentasRequest.java
@@ -1,13 +1,0 @@
-package com.test.retirarefectivo.model.registro;
-
-import lombok.Data;
-
-@Data
-public class RegistroCuentasRequest {
-
-    private String nombre;
-    private Long cuenta;
-    private Integer status;
-
-
-}
