+++ src/test/java/com/banckiko/atm/login/LoginApplicationTests.java
@@ -1,0 +1,13 @@
+package com.banckiko.atm.login;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LoginApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
