--- src/test/java/com/test/retirarefectivo/RetirarefectivoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.test.retirarefectivo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RetirarefectivoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
